Move HCPB build files to examples directory
